Bug 26295: Remove incorrect paragraph about merges requiring a reindex
Deletes:

"Important

It is important to rebuild your zebra index immediately after
merging records. If a search is performed for a record which h
as been deleted Koha will present the patrons with an error in the OPAC."

Reasoning: Every change, merge and delete requires a reindex, but
usually this is handled in the background and does not require
extra action. It is not necessary to point this out here and it's
more confusing than helpful (see discussion on bug too)