Bug 14941: [20.11] Add/correct information about substitute values in MARC bibliographic frameworks
The information on the substitute values for the Default field in
MARC biblipgrahic frameworks was incorrect and incomplete.

Note: This applies to 20.11 only, for <<YY>> being only in 20.11.
Bug 24134 - Add placeholder for 2 digit years to allow autogeneration of dates in 008

I will try to submit another patch for earlier versions.